+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 #### Improvements
 - Hybrid execution row estimate improvements and a reduction of eager calls.
 - Added support for creating permanent and immutable UDFs/UDTFs with `DataFrame/Series/GroupBy.apply`, `map`, and `transform` by passing the `snowflake_udf_params` keyword argument. See documentation for details.
+- Supported applying Snowpark functions and Snowflake Cortex functions to dataframes and series using the `Pandas` backend.
 
 #### Bug Fixes